Let widget_export_actions library to be found when gammaray is attached to a running application
The library search path is hand over to the forked process by setting
GAMMARAY_PROBE_PATH environment variable. This doesn't work when Gammaray
attaches to a running Qt application. As we know the exact location on
a Debian system, we don't have to rely on env variables and use
hardcoded GAMMARAY_LIB_INSTALL_DIR instead, which works in all cases.